# Conflicts:
#	src/Common/quoteString.cpp
#	src/Common/quoteString.h
#	src/IO/WriteHelpers.h

+++ tests/queries/0_stateless/02714_local_object_storage.sh
@@ -12,6 +12,13 @@ settings=(
     --local_filesystem_read_method='pread'
 )
 
+server_path=$(${CLICKHOUSE_CLIENT} -q "SELECT path FROM system.disks WHERE name = 'default'")
+
+if [ "${server_path:0:1}" != "/" ]; then
+    echo "path should be absolute (path: $path)" >&2
+    exit 1
+fi
+
 $CLICKHOUSE_CLIENT "${settings[@]}" -nm -q "
     DROP TABLE IF EXISTS test;
 
@@ -19,7 +26,7 @@ $CLICKHOUSE_CLIENT "${settings[@]}" -nm -q "
     ENGINE = MergeTree() ORDER BY tuple()
     SETTINGS disk = disk(
         type = 'local_blob_storage',
-        path = '${CLICKHOUSE_DISKS_FILES}/${CLICKHOUSE_TEST_UNIQUE_NAME}/');
+        path = '${CLICKHOUSE_TEST_UNIQUE_NAME}/');
 
     INSERT INTO test SELECT 1, 'test';
     SELECT * FROM test;
@@ -32,7 +39,7 @@ $CLICKHOUSE_CLIENT "${settings[@]}" -nm -q "
         type = 'cache',
         max_size = '10Mi',
         path = '${CLICKHOUSE_TEST_UNIQUE_NAME}/',
-        disk = disk(type='local_blob_storage', path='${CLICKHOUSE_DISKS_FILES}/${CLICKHOUSE_TEST_UNIQUE_NAME}/'));
+        disk = disk(type='local_blob_storage', path='${server_path}/disks/${CLICKHOUSE_TEST_UNIQUE_NAME}/'));
 
     INSERT INTO test SELECT 1, 'test';
     SELECT * FROM test;
